refactor: reorganize presenter interface structure
- Moved ShowFuturesPresenterInterface to a new file for better organization.
- Updated import paths in ShowFuturesPresenter and ShowFuturesUseCase to reflect the new structure.
