+++ src/interactor/use_cases/show_futures.py
@@ -1,6 +1,6 @@
 from src.app.cli_pfcf.config import Config
 from src.interactor.dtos.show_futures_dtos import ShowFuturesInputDto, ShowFuturesOutputDto
-from src.interactor.interfaces.show_futures_presenter_interface import ShowFuturesPresenterInterface
+from src.interactor.interfaces.presenters.show_futures_presenter import ShowFuturesPresenterInterface
 from src.interactor.interfaces.logger.logger import LoggerInterface
 from src.interactor.interfaces.repositories.session_repository import SessionRepositoryInterface
